requirements.txt
Flask==3.0.2
Pillow==10.1.0
qrcode==7.4.2
vobject==0.9.6.1 